Fully allocate instance disks at launch
By default the disk was allocated on-demand. This could lead to a
scenario where space usage was not accounted for properly and space is
exhausted if instances start filling up their local disks. Changing data
within instances could be slower too due to copy on write happening vs
just writing to the already allocated space.

util-linux is required to get the fallocate utility which is used to
allocate the space.